make sure magic calls on entity find the correct table name
I had two tables called `pages` and `page_contents`.

Beforehand when calling for example `$page->hasPageContents()` on a Page
Entity then the library looks for the database wrong table `pageContents`
because `__call` doesn't translate the method name from camelCase to
snake_case.

This commit fixes this that the snake-case table names such as
`page_contents` get found.

Beforehand when calling for example $page->hasPageContents() on a an
entity then looks for the database wroing table `page_contents`. This
commit fixes this that the snake-case table name `page_contents`
gets found.